api/requirements.txt
fastapi
fastapi_azure_auth 
uvicorn
python-dotenv
python-jose
requests
msal
